backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
sentence-transformers
python-dotenv
deep-translator
